Add dagger modifier and refactor gates into a separate file
