+++ src/main/scala/scotty/simulator/QuantumSimulator.scala
@@ -1,9 +1,9 @@
 package scotty.simulator
 
-import scotty.quantum.Gate.GateGen
 import scotty.quantum._
 import scotty.quantum.QuantumContext._
-import scotty.quantum.StandardGate
+import scotty.quantum.gate.{ControlGate, Dagger, Gate, StandardGate, SwapGate, TargetGate}
+import scotty.quantum.gate.Gate.GateGen
 import scotty.quantum.math.MathUtils
 import scotty.simulator.math.Implicits._
 
@@ -91,6 +91,7 @@ case class QuantumSimulator()(implicit random: Random = new Random) extends Quan
     case swap: SwapGate => swapMatrix(swap)
     case control: ControlGate => controlMatrix(control)
     case target: TargetGate => target.customMatrix.getOrElse(targetMatrix(target))
+    case dagger: Dagger => MatrixWrapper(dagger.target.matrix(this)).conjugateTranspose.getData
   }
 
   /**
