+++ src/test/scala/scotty/simulator/ControlledGateSpec.scala
@@ -1,8 +1,9 @@
 package scotty.simulator
 
 import org.scalatest.FlatSpec
-import scotty.quantum.StandardGate.{X, _}
+import scotty.quantum.gate.StandardGate.{X, _}
 import scotty.quantum._
+import scotty.quantum.gate.Controlled
 
 class ControlledGateSpec extends FlatSpec {
   val sim = QuantumSimulator()
